requirements.txt
neat-python
pygame~=2.5.0
